# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  media.MediaListResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      file:
        $ref: '#/definitions/storage.Attachment'
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  media.MediaResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      file:
        $ref: '#/definitions/storage.Attachment'
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  storage.Attachment:
    properties:
      created_at:
        type: string
      field:
        type: string
      filename:
        type: string
      id:
        type: integer
      model_id:
        type: integer
      model_type:
        type: string
      path:
        type: string
      size:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  types.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/types.Pagination'
    type: object
  types.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact:
    email: info@base.al
    name: Base Team
    url: https://base.al
  description: This is the API documentation for Base Framework
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://base.al/terms
  title: Base Framework API
  version: 2.0.0
paths:
...
  /media:
    get:
      description: Get a paginated list of media items
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PaginatedResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List media items
      tags:
      - Core/Media
    post:
      consumes:
      - multipart/form-data
      description: Create a new media item with optional file upload
      parameters:
      - description: Media name
        in: formData
        name: name
        required: true
        type: string
      - description: Media type
        in: formData
        name: type
        required: true
        type: string
      - description: Media description
        in: formData
        name: description
        type: string
      - description: Media file
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/media.MediaResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create a new media item
      tags:
      - Core/Media
  /media/{id}:
    delete:
      description: Delete a media item and its associated file
      parameters:
      - description: Media Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete a media item
      tags:
      - Core/Media
    get:
      description: Get a media item by Id
      parameters:
      - description: Media Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/media.MediaResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get a media item
      tags:
      - Core/Media
    put:
      consumes:
      - multipart/form-data
      description: Update a media item's details and optionally its file
      parameters:
      - description: Media Id
        in: path
        name: id
        required: true
        type: integer
      - description: Media name
        in: formData
        name: name
        type: string
      - description: Media type
        in: formData
        name: type
        type: string
      - description: Media description
        in: formData
        name: description
        type: string
      - description: Media file
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/media.MediaResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update a media item
      tags:
      - Core/Media
  /media/{id}/file:
    delete:
      description: Remove the file attached to a media item
      parameters:
      - description: Media Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/media.MediaResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Remove media file
      tags:
      - Core/Media
    put:
      consumes:
      - multipart/form-data
      description: Update the file attached to a media item
      parameters:
      - description: Media Id
        in: path
        name: id
        required: true
        type: integer
      - description: Media file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/media.MediaResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update media file
      tags:
      - Core/Media
  /media/all:
    get:
      description: Get an unpaginated list of all media items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/media.MediaListResponse'
            type: array
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List all media items
      tags:
      - Core/Media
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key for authentication
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: Enter your token with the prefix "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"